codes/requirements.txt
scikit_image==0.12.3
Django==1.10.5
numpy==1.12.0
matplotlib==2.0.0
scipy==0.18.1
Keras==2.0.4
Pillow==4.1.1
skimage==0.0
scikit_learn==0.18.1
